Slave Database Support for replicated systems
Most replication may not apply to this, however some partially
replicated Postgres servers may need to run queries directly
on replication slaves at the time of migration.

To use this, a slave connection must be specified in the command/env
and then a migration file will need a `-- migrate:up slave:true` section
containing the queries to perform.

Additionally:
- Adds tparse for nicer test output
- Fixes docker-compose configuration for latest version of docker compose module